fix: vendingMachine test DOM 참조 에러 수정
utils 파일에서 $ 함수를 사용해서 test에 에러가 발생함
snackbar.ts 라는 다른 파일로 분리하여 test내에서 DOM참조하지 않도록 수정
