+++ src/ts/snackbar.ts
@@ -1,0 +1,11 @@
+import { $ } from './utils';
+
+const $snackBar = $('#snackbar');
+
+export const alertSnackBar = (message: string) => {
+  $snackBar.classList.toggle('show');
+  $snackBar.textContent = message;
+  setTimeout(() => {
+    $snackBar.classList.toggle('show');
+  }, 3000);
+};
