+++ src/ts/utils.ts
@@ -10,18 +10,8 @@ export const $$ = <E extends Element, P extends Element>(
   baseElement: P | Document = document,
 ): NodeListOf<E> => baseElement.querySelectorAll(selector);
 
-const $snackBar = $('#snackbar');
-
 export const getRandomNumber = (min: number, max: number) => {
   return Math.floor(Math.random() * (max - min + 1)) + min;
-};
-
-export const alertSnackBar = (message: string) => {
-  $snackBar.classList.toggle('show');
-  $snackBar.textContent = message;
-  setTimeout(() => {
-    $snackBar.classList.toggle('show');
-  }, 3000);
 };
 
 export const isLogedIn = () => localStorage.getItem(STORAGE_ID.LOGIN);
